backend-services/fraud-detection-svc/requirements.txt
flask==3.0.3
requests==2.32.3
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-flask==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
